+++ container/fmi3.c
@@ -587,7 +587,7 @@ fmi3Status fmi3DoStep(fmi3Instance instance,
         ts_multiplier = 1;
 
     const double ts = container->time_step * ts_multiplier;
-    const int local_steps = ((int)((end_time - curent_time + container->tolerance) / ts)) * ts_multiplier;
+    const int local_steps = (int)((end_time - curent_time + container->tolerance) / ts);
     
     /*
      * Early return if requested end_time is lower than next container time step.